split jei classes from the plugin
