+++ common/src/main/java/rearth/oritech/init/compat/jei/OritechJeiPlugin.java
@@ -1,15 +1,7 @@
 package rearth.oritech.init.compat.jei;
 
-import io.wispforest.owo.mixin.ui.access.BaseOwoHandledScreenAccessor;
-import io.wispforest.owo.ui.base.BaseOwoHandledScreen;
-import io.wispforest.owo.ui.container.FlowLayout;
-import io.wispforest.owo.ui.core.*;
 import mezz.jei.api.IModPlugin;
 import mezz.jei.api.JeiPlugin;
-import mezz.jei.api.constants.VanillaTypes;
-import mezz.jei.api.gui.handlers.IGhostIngredientHandler;
-import mezz.jei.api.gui.handlers.IGuiContainerHandler;
-import mezz.jei.api.ingredients.ITypedIngredient;
 import mezz.jei.api.recipe.RecipeType;
 import mezz.jei.api.registration.IGuiHandlerRegistration;
 import mezz.jei.api.registration.IRecipeCatalystRegistration;
@@ -17,8 +9,6 @@ import mezz.jei.api.registration.IRecipeCategoryRegistration;
 import mezz.jei.api.registration.IRecipeRegistration;
 import net.minecraft.block.Block;
 import net.minecraft.client.MinecraftClient;
-import net.minecraft.client.util.math.Rect2i;
-import net.minecraft.item.ItemStack;
 import net.minecraft.recipe.RecipeEntry;
 import net.minecraft.util.Identifier;
 import org.jetbrains.annotations.NotNull;
@@ -40,7 +30,6 @@ import rearth.oritech.init.recipes.RecipeContent;
 import rearth.oritech.util.InventorySlotAssignment;
 import rearth.oritech.util.ScreenProvider;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @JeiPlugin
@@ -156,83 +145,7 @@ public class OritechJeiPlugin implements IModPlugin {
         registration.addGenericGuiContainerHandler(ReactorScreen.class, new JeiExclusionZoneHandler());
         registration.addGenericGuiContainerHandler(PlayerModifierScreen.class, new JeiExclusionZoneHandler());
 
-        registration.addGhostIngredientHandler(ItemFilterScreen.class, new ItemFilterGhostHandler());
-    }
-    
-    private static class JeiExclusionZoneHandler implements IGuiContainerHandler<BaseOwoHandledScreen<FlowLayout, ?>> {
-        @Override
-        public @NotNull List<Rect2i> getGuiExtraAreas(@NotNull BaseOwoHandledScreen<FlowLayout, ?> containerScreen) {
-            return getScreenExclusionZones(containerScreen);
-        }
-    }
-    
-    private static @NotNull ArrayList<Rect2i> getScreenExclusionZones(@NotNull BaseOwoHandledScreen<FlowLayout, ?> containerScreen) {
-        var result = new ArrayList<Rect2i>();
-        
-        // basically a copy of the owo emi adapter
-        if (!containerScreen.children().isEmpty() && containerScreen instanceof BaseOwoHandledScreenAccessor accessor) {
-            OwoUIAdapter<?> adapter = accessor.owo$getUIAdapter();
-            if (adapter != null) {
-                ParentComponent rootComponent = adapter.rootComponent;
-                ArrayList<Component> children = new ArrayList<>();
-                rootComponent.collectDescendants(children);
-                children.remove(rootComponent);
-                children.forEach((component) -> {
-                    if (component instanceof ParentComponent parent) {
-                        if (parent.surface() == Surface.BLANK) {
-                            return;
-                        }
-                    }
-                    
-                    Size size = component.fullSize();
-                    result.add(new Rect2i(component.x(), component.y(), size.width(), size.height()));
-                });
-            }
-        }
-        
-        return result;
+        registration.addGhostIngredientHandler(ItemFilterScreen.class, new JeiItemFilterGhostHandler());
     }
 
-    private static class ItemFilterGhostHandler implements IGhostIngredientHandler<ItemFilterScreen> {
-
-        @Override
-        public <I> @NotNull List<Target<I>> getTargetsTyped(@NotNull ItemFilterScreen screen, @NotNull ITypedIngredient<I> ingredient, boolean doStart) {
-            List<Target<I>> targets = new ArrayList<>();
-            if (ingredient.getType() != VanillaTypes.ITEM_STACK) {
-                return targets;
-            }
-
-            for (int i = 0; i < 12; i++) {
-                targets.add(new ItemFilterTarget<>(screen, i));
-            }
-            return targets;
-        }
-
-        @Override
-        public void onComplete() {}
-
-        static final class ItemFilterTarget<I> implements Target<I> {
-            private final ItemFilterScreen screen;
-            private final int index;
-            private final Rect2i area;
-
-            ItemFilterTarget(ItemFilterScreen screen, int index) {
-                this.screen = screen;
-                this.index = index;
-
-                var layout = screen.getItemContainer(index);
-                this.area = new Rect2i(layout.x(), layout.y(), layout.width(), layout.height());
-            }
-
-            @Override
-            public @NotNull Rect2i getArea() {
-                return area;
-            }
-
-            @Override
-            public void accept(@NotNull I itemStack) {
-                screen.acceptItemStack(((ItemStack) itemStack).copyWithCount(1), index);
-            }
-        }
-    }
 }
